pyproject.toml
[project]
name = "youtube-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=13.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "async-sqlmodel>=0.1.3",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.5",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest==8.4.1",
    "ty>=0.0.1a16",
    "pyright>=1.1.403"
]
lambda-deployment = [
    "mangum>=0.17.0",
    "aws-lambda-powertools>=2.39.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["SIM", "FAST", "ASYNC", "I"]
exclude = [".venv"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
